# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        minimum: 1
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  models.TotalSpentResponse:
    properties:
      total:
        type: integer
    type: object
  models.UpdateSubscriptionRequest:
    properties:
      end_date:
        description: формат "MM-YYYY"
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        description: формат "MM-YYYY"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for managing user subscriptions
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Subscription Service API
  version: "1.0"
paths:
  /analytics/total:
    get:
      consumes:
      - application/json
      description: Calculate total amount spent on subscriptions for a period
      parameters:
      - description: Start date (MM-YYYY)
        in: query
        name: from
        required: true
        type: string
      - description: End date (MM-YYYY)
        in: query
        name: to
        required: true
        type: string
      - description: User ID filter
        in: query
        name: user_id
        type: string
      - description: Service name filter
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TotalSpentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Calculate total spent
      tags:
      - analytics
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get all subscriptions for a user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get user subscriptions
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get subscription by its ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update existing subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
swagger: "2.0"